// src/app/models/iteration.ts
export class Iteration {
    id: number;
    name: string;
    number: string;
    increment_id: number;
    increment_name: number;
    increment_number: number;
    status_id: number;
    status_name: string;
    start_date: string;
    end_date: string;
    story_points: number;
    completeness: number;
}